+++ praw/models/reddit/comment.py
@@ -1,5 +1,4 @@
 """Provide the Comment class."""
-from ...const import urljoin
 from ...exceptions import ClientException
 from ..comment_forest import CommentForest
 from .base import RedditBase
@@ -75,6 +74,8 @@ class Comment(RedditBase, InboxableMixin, UserContentMixin):
             attribute = '_replies'
         elif attribute == 'subreddit':
             value = self._reddit.subreddit(value)
+        elif attribute == 'permalink':
+            attribute = '_permalink'  # remove these two lines in PRAW 6
         super(Comment, self).__setattr__(attribute, value)
 
     def _extract_submission_id(self):
@@ -149,23 +150,17 @@ class Comment(RedditBase, InboxableMixin, UserContentMixin):
     def permalink(self, fast=False):
         """Return a permalink to the comment.
 
-        :param fast: Return the result as quickly as possible (default: False).
+        :param fast: Has no effect. Accepted to maintain backwards
+            compatibility.
 
-        In order to determine the full permalink for a comment, the Submission
-        may need to be fetched if it hasn't been already. Set ``fast=True`` if
-        you want to bypass that possible load.
-
-        A full permalink looks like:
-        /r/redditdev/comments/2gmzqe/praw_https_enabled/cklhv0f
-
-        A fast-loaded permalink for the same comment will look like:
-        /comments/2gmzqe//cklhv0f
+        .. warning:: (Deprecated) This method will be removed in PRAW 6 because
+            permalink is now a first class attribute.
 
         """
-        # pylint: disable=no-member
-        if not fast or 'permalink' in self.submission.__dict__:
-            return urljoin(self.submission.permalink, self.id)
-        return '/comments/{}//{}'.format(self.submission.id, self.id)
+        # pylint: disable=unused-argument
+        if not self._fetched:
+            self._fetch()  # old way required fetching self as well
+        return self._permalink
 
     def refresh(self):
         """Refresh the comment's attributes.