+++ CHANGES.rst
@@ -13,6 +13,8 @@ Unreleased
 
 * ``subreddit.comments.gilded`` because there isn't actually an endpoint that
   returns only gilded comments. Use ``subreddit.gilded`` instead.
+* ``comment.permalink`` because permalink is now an attribute returned by
+  Reddit.
 
 
 5.1.0 (2017/08/31)